refactor: consolidate SFTP file deletion logic
- Moved file deletion suffix logic from a separate module to a static method within the `SFTPStorage` class, enhancing encapsulation and reducing module dependencies.
- Updated relevant methods to utilize the new static method for generating delete path suffixes, ensuring consistency across file operations.
- Removed the now-unnecessary `fileStorageDeleteSuffixes.ts` file to streamline the codebase.
